requirements.txt
streamlit==1.35.0
requests==2.31.0
python-dotenv
google-generativeai
pdfplumber
Pillow
pytesseract
pdf2image
poppler-utils   # (Linux)


